requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
beautifulsoup4==4.13.4
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.8
colorama==0.4.6
comm==0.2.3
curl_cffi==0.13.0
debugpy==1.8.16
decorator==5.2.1
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.116.1
frozendict==2.4.6
h11==0.16.0
httptools==0.6.4
idna==3.10
importlib_metadata==8.7.0
ipykernel==6.30.1
ipython==8.18.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.8.1
lxml==6.0.0
matplotlib-inline==0.1.7
multitasking==0.0.12
nest-asyncio==1.6.0
newsapi-python==0.2.7
numpy==2.0.2
packaging==25.0
pandas==2.3.1
parso==0.8.4
peewee==3.18.2
platformdirs==4.3.8
prompt_toolkit==3.0.51
protobuf==6.32.0
psutil==7.0.0
psycopg2-binary==2.9.10
pure_eval==0.2.3
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
pyrate-limiter==3.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==311
PyYAML==6.0.2
pyzmq==27.0.1
requests==2.32.5
sec-edgar-api==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
stack-data==0.6.3
starlette==0.47.2
tornado==6.5.2
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
yfinance==0.2.65
zipp==3.23.0
